Switch to nunjucks and use front-matter
Unfortunately swig is no longer supported. Fortunately nunjucks is based
on very similar syntax and so I have switched to nunjucks.

The uses of yaml front-matter is now also somewhat common and seems like a
more elegant solution than the directory structure I had before so I've
moved the metadata for the lone example post into the template as front
matter.

Also renamed the site css to remove the reference to the old domain name.